+++ rollup.config.js
@@ -1,9 +1,7 @@
 import resolve from "@rollup/plugin-node-resolve"
 import typescript from "@rollup/plugin-typescript"
-import { terser } from "rollup-plugin-terser"
-import commonjs from "@rollup/plugin-commonjs"
-import json from "@rollup/plugin-json"
-import pkg from "./package.json"
+import terser from "@rollup/plugin-terser"
+import pkg from "./package.json" assert { type: "json" }
 
 export default {
 	input: "src/index.ts",
@@ -13,8 +11,6 @@ export default {
 	],
 	plugins: [
 		typescript(),
-		commonjs(),
-		json(),
 		resolve(),
 		terser({
 			format: {
@@ -22,5 +18,5 @@ export default {
 			}
 		})
 	],
-	external: ["env-var", "universal-cookie", "njwt", "logical-cas-client"]
+	external: Object.keys(pkg.dependencies)
 }