+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0]
+
+### Changed
+
+- Add compatibility to all SvelteKit version `1.0.0-next.492` and later
+- Use [@macfja/sveltekit-session](https://www.npmjs.com/package/@macfja/sveltekit-session) to data persistence
+
 ## [1.0.1]
 
 ### Fixed
@@ -18,5 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 First version
 
-[unreleased]: https://github.com/MacFJA/sveltekit-cas/compare/1.0.0...HEAD
+[unreleased]: https://github.com/MacFJA/sveltekit-cas/compare/2.0.0...HEAD
+[2.0.0]: https://github.com/MacFJA/sveltekit-cas/releases/tag/2.0.0
+[1.0.1]: https://github.com/MacFJA/sveltekit-cas/releases/tag/1.0.1
 [1.0.0]: https://github.com/MacFJA/sveltekit-cas/releases/tag/1.0.0